+++ contracts/token/oft/extension/PausableOFT.sol
@@ -7,18 +7,13 @@ import "@openzeppelin/contracts/security/Pausable.sol";
 
 // allow OFT to pause all cross-chain transactions
 contract PausableOFT is OFT, Pausable {
-    constructor(
-        string memory _name,
-        string memory _symbol,
-        address _lzEndpoint,
-        uint256 _initialSupply
-    ) OFT(_name, _symbol, _lzEndpoint, _initialSupply) {}
+    constructor(string memory _name, string memory _symbol, address _lzEndpoint, uint _initialSupply) OFT(_name, _symbol, _lzEndpoint, _initialSupply) {}
 
     function _debitFrom(
         address _from,
         uint16, // _dstChainId
         bytes memory, // _toAddress
-        uint256 _amount
+        uint _amount
     ) internal override whenNotPaused {
         _burn(_from, _amount);
     }
